Restore old types for ObservableStream
